requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.14.3
boto3==1.42.54
botocore==1.42.54
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
curl_cffi==0.13.0
filelock==3.24.3
frozendict==2.4.7
idna==3.11
jmespath==1.1.0
multitasking==0.0.12
numpy==2.4.2
packaging==26.0
pandas==3.0.1
peewee==4.0.0
platformdirs==4.9.2
protobuf==6.33.5
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==4.3.0
sortedcontainers==2.4.0
soupsieve==2.8.3
tomlkit==0.14.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
